Convert storage-users to use service tracing.
This converts the storage users service to use the service tracing provider.
