+++ services/storage-users/pkg/revaconfig/config.go
@@ -11,11 +11,6 @@ import (
 func StorageUsersConfigFromStruct(cfg *config.Config) map[string]interface{} {
 	rcfg := map[string]interface{}{
 		"core": map[string]interface{}{
-			"tracing_enabled":           cfg.Tracing.Enabled,
-			"tracing_exporter":          cfg.Tracing.Type,
-			"tracing_endpoint":          cfg.Tracing.Endpoint,
-			"tracing_collector":         cfg.Tracing.Collector,
-			"tracing_service_name":      cfg.Service.Name,
 			"graceful_shutdown_timeout": cfg.GracefulShutdownTimeout,
 		},
 		"shared": map[string]interface{}{
